auto-claude: subtask-3-1 - Extract SettingsModal component
