+++ .auto-claude-status
@@ -3,23 +3,23 @@
   "spec": "003-split-monolithic-app-tsx-into-smaller-components",
   "state": "building",
   "subtasks": {
-    "completed": 3,
+    "completed": 5,
     "total": 14,
     "in_progress": 1,
     "failed": 0
   },
   "phase": {
-    "current": "Extract Custom Hooks",
+    "current": "Extract Modal Components",
     "id": null,
-    "total": 3
+    "total": 2
   },
   "workers": {
     "active": 0,
     "max": 1
   },
   "session": {
-    "number": 5,
+    "number": 7,
     "started_at": "2026-01-31T18:01:49.917579"
   },
-  "last_update": "2026-01-31T18:11:09.737999"
+  "last_update": "2026-01-31T18:15:30.426500"
 }
